requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.0.0
alembic>=1.10.0

# Data Processing
pandas>=2.0.0

# Agent Framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0

# AI/ML
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0
vertexai>=1.60.0
openai>=1.40.0

# API Integrations
amadeus>=8.0.0
requests>=2.30.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.0.0
mypy>=1.7.0

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.19.0